Add execution env hooks to separate files
