+++ frontend/awx/administration/execution-environments/ExecutionEnvironments.tsx
@@ -1,31 +1,5 @@
-import { ButtonVariant } from '@patternfly/react-core';
-import { PencilAltIcon, PlusIcon, TrashIcon } from '@patternfly/react-icons';
-import { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  IPageAction,
-  ITableColumn,
-  IToolbarFilter,
-  PageActionSelection,
-  PageActionType,
-  PageHeader,
-  PageLayout,
-  PageTable,
-  usePageNavigate,
-} from '../../../../framework';
-import {
-  useCreatedColumn,
-  useDescriptionColumn,
-  useIdColumn,
-  useModifiedColumn,
-  useNameColumn,
-  useOrganizationNameColumn,
-} from '../../../common/columns';
-import {
-  useImageToolbarFilter,
-  useNameToolbarFilter,
-  useOrganizationToolbarFilter,
-} from '../../common/awx-toolbar-filters';
+import { PageHeader, PageLayout, PageTable, usePageNavigate } from '../../../../framework';
 import { useAwxView } from '../../common/useAwxView';
 import { ExecutionEnvironment } from '../../interfaces/ExecutionEnvironment';
 
@@ -33,14 +7,10 @@ import { awxAPI } from '../../common/api/awx-utils';
 import { useAwxConfig } from '../../common/useAwxConfig';
 import { getDocsBaseUrl } from '../../common/util/getDocsBaseUrl';
 import { AwxRoute } from '../../main/AwxRoutes';
-import { useDeleteExecutionEnvironments } from './hooks/useDeleteExecutionEnvironments';
-import {
-  cannotDeleteResource,
-  cannotDeleteResources,
-  cannotEditResource,
-} from '../../../common/utils/RBAChelpers';
-import { useOptions } from '../../../common/crud/useOptions';
-import { OptionsResponse, ActionsResponse } from '../../interfaces/OptionsResponse';
+import { useExecutionEnvRowActions } from './hooks/useExecutionEnvRowActions';
+import { useExecutionEnvToolbarActions } from './hooks/useExecutionEnvToolbarActions';
+import { useExecutionEnvironmentsColumns } from './hooks/useExecutionEnvironmentsColumns';
+import { useExecutionEnvironmentsFilters } from './hooks/useExecutionEnvironmentsFilters';
 
 export function ExecutionEnvironments() {
   const { t } = useTranslation();
@@ -53,68 +23,9 @@ export function ExecutionEnvironments() {
     toolbarFilters,
     tableColumns,
   });
-  const deleteExecutionEnvironments = useDeleteExecutionEnvironments(view.unselectItemsAndRefresh);
 
-  const { data } = useOptions<OptionsResponse<ActionsResponse>>(awxAPI`/execution_environments/`);
-  const canCreateExecutionEnvironment = Boolean(data && data.actions && data.actions['POST']);
-
-  const toolbarActions = useMemo<IPageAction<ExecutionEnvironment>[]>(
-    () => [
-      {
-        type: PageActionType.Button,
-        selection: PageActionSelection.None,
-        variant: ButtonVariant.primary,
-        isPinned: true,
-        icon: PlusIcon,
-        label: t('Create execution environment'),
-        isDisabled: canCreateExecutionEnvironment
-          ? undefined
-          : t(
-              'You do not have permission to create an execution environment. Please contact your organization administrator if there is an issue with your access.'
-            ),
-        onClick: () => pageNavigate(AwxRoute.CreateExecutionEnvironment),
-      },
-      { type: PageActionType.Seperator },
-      {
-        type: PageActionType.Button,
-        selection: PageActionSelection.Multiple,
-        icon: TrashIcon,
-        label: t('Delete selected execution environments'),
-        isDisabled: (executionEnvironments) => cannotDeleteResources(executionEnvironments, t),
-        onClick: deleteExecutionEnvironments,
-        isDanger: true,
-      },
-    ],
-    [t, canCreateExecutionEnvironment, deleteExecutionEnvironments, pageNavigate]
-  );
-
-  const rowActions = useMemo<IPageAction<ExecutionEnvironment>[]>(
-    () => [
-      {
-        type: PageActionType.Button,
-        selection: PageActionSelection.Single,
-        icon: PencilAltIcon,
-        isPinned: true,
-        label: t('Edit execution environment'),
-        isDisabled: (executionEnvironment) => cannotEditResource(executionEnvironment, t),
-        onClick: (executionEnvironment) =>
-          pageNavigate(AwxRoute.EditExecutionEnvironment, {
-            params: { id: executionEnvironment.id },
-          }),
-      },
-      { type: PageActionType.Seperator },
-      {
-        type: PageActionType.Button,
-        selection: PageActionSelection.Single,
-        icon: TrashIcon,
-        label: t('Delete execution environment'),
-        isDisabled: (executionEnvironment) => cannotDeleteResource(executionEnvironment, t),
-        onClick: (executionEnvironment) => deleteExecutionEnvironments([executionEnvironment]),
-        isDanger: true,
-      },
-    ],
-    [pageNavigate, deleteExecutionEnvironments, t]
-  );
+  const rowActions = useExecutionEnvRowActions(view);
+  const toolbarActions = useExecutionEnvToolbarActions(view);
 
   return (
     <PageLayout>
@@ -145,60 +56,4 @@ export function ExecutionEnvironments() {
       />
     </PageLayout>
   );
-}
-
-export function useExecutionEnvironmentsFilters() {
-  const nameToolbarFilter = useNameToolbarFilter();
-  const organizationToolbarFilter = useOrganizationToolbarFilter();
-  const imageToolbarFilter = useImageToolbarFilter();
-  const toolbarFilters = useMemo<IToolbarFilter[]>(
-    () => [nameToolbarFilter, organizationToolbarFilter, imageToolbarFilter],
-    [nameToolbarFilter, organizationToolbarFilter, imageToolbarFilter]
-  );
-  return toolbarFilters;
-}
-
-export function useExecutionEnvironmentsColumns(options?: {
-  disableSort?: boolean;
-  disableLinks?: boolean;
-}) {
-  const { t } = useTranslation();
-  const pageNavigate = usePageNavigate();
-  const nameClick = useCallback(
-    (executionEnvironment: ExecutionEnvironment) =>
-      pageNavigate(AwxRoute.ExecutionEnvironmentDetails, {
-        params: { id: executionEnvironment.id },
-      }),
-    [pageNavigate]
-  );
-  const nameColumn = useNameColumn({
-    ...options,
-    onClick: nameClick,
-  });
-  const idColumn = useIdColumn<ExecutionEnvironment>();
-  const descriptionColumn = useDescriptionColumn();
-  const organizationColumn = useOrganizationNameColumn(
-    AwxRoute.OrganizationDetails,
-    options,
-    t('Globally available')
-  );
-  const createdColumn = useCreatedColumn(options);
-  const modifiedColumn = useModifiedColumn(options);
-  const tableColumns = useMemo<ITableColumn<ExecutionEnvironment>[]>(
-    () => [
-      idColumn,
-      nameColumn,
-      descriptionColumn,
-      {
-        header: t('Image'),
-        cell: (executionEnvironment) => executionEnvironment.image,
-        sort: 'image',
-      },
-      organizationColumn,
-      createdColumn,
-      modifiedColumn,
-    ],
-    [idColumn, nameColumn, descriptionColumn, t, organizationColumn, createdColumn, modifiedColumn]
-  );
-  return tableColumns;
 }
